applied pr review suggestions
